refactor FastjsonSockJsMessageCodec package and testcase.
