+++ extension/src/test/java/com/alibaba/fastjson2/spring/FastjsonSockJsMessageCodecTest.java
@@ -1,11 +1,11 @@
 package com.alibaba.fastjson2.spring;
 
-import com.alibaba.fastjson2.support.spring.FastjsonSockJsMessageCodec;
+import com.alibaba.fastjson2.support.spring.websocket.sockjs.FastjsonSockJsMessageCodec;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class FastjsonSockJsMessageCodecTest_encode {
+public class FastjsonSockJsMessageCodecTest {
     @Test
     public void test_encode() {
         FastjsonSockJsMessageCodec fastjsonCodec = new FastjsonSockJsMessageCodec();
